# generated/src/main/java/com/api/TalentTorokuKoshinBFFApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.api;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.model.MTalent;
import com.model.TalentTorokuKoshinBFF;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-11-04T23:34:19.826482+09:00[Asia/Tokyo]")
@Validated
@Tag(name = "talentTorokuKoshinBFF", description = "the talentTorokuKoshinBFF API")
@RequestMapping("api")
public interface TalentTorokuKoshinBFFApi {

    /**
     * POST /talentTorokuKoshinBFF : タレント情報を登録または更新するBFF
     * タレント情報を登録または更新するBFF
     *
     * @param mtalent タレント情報を登録または更新するBFF (optional)
     * @return 登録または更新したタレントマスタ情報 (status code 200)
     */
    @Operation(
        operationId = "postTalentTorokuKoshinBFF",
        summary = "タレント情報を登録または更新するBFF",
        tags = { "talentTorokuKoshinBFF" },
        responses = {
            @ApiResponse(responseCode = "200", description = "登録または更新したタレントマスタ情報", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TalentTorokuKoshinBFF.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/talentTorokuKoshinBFF",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<TalentTorokuKoshinBFF> postTalentTorokuKoshinBFF(
        @Parameter(name = "MTalent", description = "タレント情報を登録または更新するBFF") @Valid @RequestBody(required = false) MTalent mtalent
    );

}