# openapi/talentInfoBFF.yaml
openapi: 3.0.0
info:
  title: タレント情報BFF
  description: タレントIDによってタレント情報を返すBFF
  version: 1.0.0
servers:
  - url: https://localhost:8081
    description: タレント情報BFF
paths:
  /talentInfoBFF/{talentId}:
    get:
      tags:
        - talentInfoBFF
      operationId: getTalentInfoBFF
      summary: タレント情報をタレントIDによって取得するBFF
      description: タレント情報をタレントIDによって取得するBFF
      parameters:
        - in: path
          name: talentId
          required: true
          schema:
            type: string
            maxLength: 8
      responses:
        '200':
          description: タレント情報を正常取得
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/talentInfo'
  /talentTorokuKoshinBFF:
    post:
      tags:
        - talentTorokuKoshinBFF
      operationId: postTalentTorokuKoshinBFF
      summary: タレント情報を登録または更新するBFF
      description: タレント情報を登録または更新するBFF
      requestBody:
        description: タレント情報を登録または更新するBFF
        content:
          application/json:
            schema:
              $ref: "../../TalentSearchCommon/Tables/mTalent.yaml#/components/schemas/mTalent"
      responses:
        '200':
          description: 登録または更新したタレントマスタ情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/talentTorokuKoshinBFF'
components:
  schemas:
    talentInfo:
      description: タレント情報を正常取得
      type: object
      properties:
        talentId:
          type: string
          maxLength: 8
          description: タレントID
        talentName:
          type: string
          maxLength: 30
          description: タレント名
        genreId:
          type: integer
          maximum : 000
          minimum : 999
          description: ジャンルID
        genre:
          type: string
          maxLength: 30
          description: ジャンル名
    TalentList:
      description: タレントマスタDTOの情報
      type: object
      properties:
        mTalent:
          description: タレントマスタDTO
          type: array
          items:
           $ref: "../../TalentSearchCommon/Tables/mTalent.yaml#/components/schemas/mTalent"
    talentTorokuKoshinBFF:
      description: 登録または更新したタレントマスタ情報
      type: object
      properties:
        mTalent:
          description: タレントマスタDTO
          type: array
          items:
            $ref: "../../TalentSearchCommon/Tables/mTalent.yaml#/components/schemas/mTalent"